tests added

+++ main-server/src/main/java/ru/practicum/ewm/event/model/dto/LocationDto.java
@@ -10,7 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class LocationDto {
-
-    private double latitude;
-    private double longitude;
+    private float latitude;
+    private float longitude;
 }
